requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.1
future==0.18.3
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyserial==3.5
PyYAML==6.0
Werkzeug==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
gunicorn==21.2.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1